+++ components/youtube_data_api/sources/new-subscriber/new-subscriber.mjs
@@ -1,55 +1,16 @@
-import common from "../common.mjs";
+import youtubeDataApi from "../../youtube_data_api.app.mjs";
+import common from "./common.mjs";
 
 export default {
+  ...common,
   type: "source",
   key: "youtube_data_api-new-subscriber",
   name: "New Subscriber",
   description: "Emit new event for each new Youtube subscriber to user Channel.",
-  version: "0.0.1",
+  version: "0.0.2",
   dedupe: "unique",
   props: {
+    youtubeDataApi,
     ...common.props,
-  },
-  methods: {
-    _getLastExecutionDate() {
-      return this.db.get("lastExecutionDate");
-    },
-    _setLastExecutionDate(lastExecutionDate) {
-      this.db.set("lastExecutionDate", lastExecutionDate);
-    },
-    async _emitLastSubscriptions() {
-      const lastExecutionDate = this._getLastExecutionDate();
-      let nextPageToken = null;
-      do {
-        const res = await this.youtubeDataApi.getSubscriptions({
-          part: "id,subscriberSnippet,snippet",
-          mySubscribers: true,
-          maxResults: 50,
-          pageToken: nextPageToken,
-        });
-
-        for (const item of res.data.items) {
-          if (new Date(item.snippet.publishedAt).getTime() < lastExecutionDate) {
-            return;
-          }
-          this.$emit(item, {
-            id: item.id,
-            summary: item.subscriberSnippet.title,
-            ts: new Date(item.snippet.publishedAt),
-          });
-        }
-
-        nextPageToken = res.data.nextPageToken;
-      } while (nextPageToken);
-    },
-  },
-  hooks: {
-    activate() {
-      this._setLastExecutionDate(new Date().getTime());
-    },
-  },
-  async run() {
-    await this._emitLastSubscriptions();
-    this._setLastExecutionDate(new Date().getTime());
   },
 };
