+++ components/youtube_data_api/sources/new-videos-by-search/common.mjs
@@ -1,0 +1,26 @@
+import common from "../common.mjs";
+
+export default {
+  ...common,
+  props: {
+    ...common.props,
+    q: {
+      type: "string",
+      label: "Search Query",
+      description: "Search for new videos that match these keywords.",
+    },
+  },
+  hooks: {
+    ...common.hooks,
+    deploy() {},
+  },
+  methods: {
+    ...common.methods,
+    getParams() {
+      return {
+        q: this.q,
+        maxResults: this.maxResults,
+      };
+    },
+  },
+};
