+++ components/youtube_data_api/sources/new-videos/common.mjs
@@ -1,0 +1,25 @@
+import common from "../common.mjs";
+
+export default {
+  ...common,
+  hooks: {
+    ...common.hooks,
+    deploy() {},
+  },
+  methods: {
+    ...common.methods,
+    getParams() {
+      return {
+        forMine: true,
+        maxResults: this.maxResults,
+      };
+    },
+  },
+  async run() {
+    const params = {
+      ...this._getBaseParams(),
+      ...this.getParams(),
+    };
+    await this.paginateVideos(params);
+  },
+};
