+++ components/youtube_data_api_custom_app/sources/new-liked-videos/new-liked-videos.mjs
@@ -1,0 +1,22 @@
+import youtubeDataApi from "../../youtube_data_api_custom_app.app.mjs";
+import common from "../../../youtube_data_api/sources/new-liked-videos/common.mjs";
+
+export default {
+  ...common,
+  key: "youtube_data_api_custom_app-new-liked-videos",
+  name: "New Liked Videos",
+  description: "Emit new event for each new Youtube video liked by the authenticated user",
+  version: "0.0.1",
+  type: "source",
+  dedupe: "unique",
+  props: {
+    youtubeDataApi,
+    maxResults: {
+      propDefinition: [
+        youtubeDataApi,
+        "maxResults",
+      ],
+    },
+    ...common.props,
+  },
+};
