+++ components/youtube_data_api_custom_app/sources/new-videos-by-search/new-videos-by-search.mjs
@@ -1,13 +1,13 @@
-import youtubeDataApi from "../../youtube_data_api.app.mjs";
-import common from "./common.mjs";
+import youtubeDataApi from "../../youtube_data_api_custom_app.app.mjs";
+import common from "../../../youtube_data_api/sources/new-videos-by-search/common.mjs";
 
 export default {
   ...common,
-  type: "source",
-  key: "youtube_data_api-new-videos-by-search",
+  key: "youtube_data_api_custom_app-new-videos-by-search",
   name: "New Videos by Search",
   description: "Emit new event for each new YouTube video matching a search query.",
-  version: "0.0.7",
+  version: "0.0.1",
+  type: "source",
   dedupe: "unique",
   props: {
     youtubeDataApi,
